Separated widget into submodule
